Fix example filtering logic w.r.t. targets
Since we compute the length of the targets tensor as
`max_target_seq_length * max_docs_per_packed_example`, it's possible
for this to end up being longer than the inputs length
(`max_docs_per_packed_example * chunk_length * chunks_per_doc`).

If this ends up being the case, then we will (currently) filter out all
examples - this PR adds a method for our PackedProblem to override to
specify a different length to condition on for the `targets`.